ffi: move two event type enums over to matrix-sdk-ffi/event.rs
